devShell: remove usage of external devShell, implement own module system
This is because numtide's devShell (although being awesome) does not use
nixpkgs' mkShell function, and hence cannot automatically provide the
dependencies of the top developpers are currently editing. This means
that developers would have to specify each depency for both the top
build and the devShell.

Here with mkShell, we can just use `inputsFrom`