lib: refactor evaluation, make a function pkgs -> eval
this moves the evaluation of the EPNix configuration into its own file

this adds a new `epnixFunEval`
which is a function that takes `pkgs` as an argument,
and evals the EPNix config with this specific package set

this function will be used so that
an EPNix support module can be used by another EPNix top.

It will reuse the same dependencies
(the same epics-base, the same support modules, the same system libraries),
and will reuse the same nixpkgs configuration
(for example cross-compilation configuration).