docs/ioc: expand Integration tests tutorial
Renamed to "Adding integration tests to your IOC", because an
"Integration tests" user guide is planned, and to make it a bit clearer
that it is a continuation of the StreamDevice tutorial.

Previously the tutorial was a mix between a guide and tutorial, not
really guided enough.

Now the tutorial builds upon the StreamDevice tutorial, and shows a
complete path to have a fully functional integration test.

Also introduce how to *run* tests, both non-interactively and
interactively.

Added a quick reference of useful Python test functions.

The tutorial was also updated due to some changes in the default
template, and some renamed options.

(cherry picked from commit b46224ce69fdbf2dfb8cbcbdcf01db7e692f34bd)