phoebus-*: use maven.test.skip instead of skipTests
Using this flag instead skips test compilation, which reduces build times by about 25% on my machine. The build times for phoebus get kind of long on our gitlab runners (~15 minutes), so reducing it by a significant amount may be helpful.

Documentation does say that it is not recommended to use this flag, but especially for the phoebus-deps package, it seems unnecessary to compile the tests, when they are not even going to be ran.

See also [the documentation](https://maven.apache.org/surefire/maven-surefire-plugin/test-mojo.html#skip) and [this SO post](https://stackoverflow.com/questions/2593588/maven-skip-building-test-classes).