pkgs/docs: create "alias" of the book package
The "docs" package is basically the same as the "book" package,
but it builds the doc in "share/doc/epnix/html",
as expected by the "build-docs-multiversion" script on master.

This allows us to build this older version of the documentation,
and make it show up on the EPNix website.

(cherry picked from commit 2da9df180b11c5e8a63732d0d89ad80d9e36b24f)