build-docs-multiversion: fetch EPNix versions by itself
The CI script run by Nix now downloads the EPNix archives by itself,
instead of requiring the GitHub workflow to do it.

This simplifies the workflow, and enables us to just run:

    nix run .#ci-scripts/build-docs-multiversion

Which will do the whole EPNix website build.

This is also one less place to edit for new EPNix releases.